change bedpe filtering script to handle both bedpe and bed regions files, and renamed the python script to better reflect its purpose
